Refactored new code into separate files for better organization
