+++ original/eval/main.py
@@ -5,7 +5,7 @@ from eval.stats import StatFile, DEFAULT_INCLUDED_STATS
 from eval.signals_dict import SignalsDict
 from eval.files import get_raw_signals_dict
 from eval.reference_stats import get_reference_data
-
+from eval.correlation import rank_correlation
 
 def get_csv_lines_from_files(
     signals_dict: SignalsDict,
@@ -27,73 +27,6 @@ def get_csv_lines_from_files(
         csv_lines.append(line)
 
     return csv_lines
-
-#TODO: move this somewhere else
-from scipy.stats import spearmanr, pearson3
-from eval import INDIVIDUAL_CHANS
-import pandas as pd
-
-
-def split_channels_from_list(ch_list: str):
-    remaining = ch_list
-    channels = []
-    updated = True # needs to be True to start
-
-    # this isn't the most efficient overall, but the search domain is small so it works.  If it needs optimization
-    # we can do it later. It just looks at signals and checks if the "remaining" starts with a known signal.  It
-    # removes it from "remaining" if it does, then tries again until remaining is len(0)
-    while len(remaining) > 0 and updated:
-        updated = False
-        for name_to_search in INDIVIDUAL_CHANS:
-            if remaining.startswith(name_to_search):
-                channels.append(name_to_search)
-                remaining = remaining[len(name_to_search):]
-                updated = True
-
-    if len(remaining) > 0:
-        raise Exception(f"Invalid channel list.  Contains unknown channels: {remaining}")
-
-    return channels
-
-def dataframe_from_signals(signals_dict: SignalsDict) -> pd.DataFrame:
-    all_signals = INDIVIDUAL_CHANS
-
-    data = {}
-    for signals, stats_dict in signals_dict.items():
-        present_signals = split_channels_from_list(signals)
-
-        # set a 1 in the signal row if the signal is used or 0 otherwise
-        for s in all_signals:
-            data[s] = data.get(s, []) + [1 if s in present_signals else 0]
-
-        for stat, value in stats_dict.items():
-            data[stat] = data.get(stat, []) + [value.mean]
-    return pd.DataFrame(data)
-
-def rank_correlation(signals_dict: SignalsDict, statistics=['F1', 'AUROC']) -> dict[str, dict[str, float]]:
-    # Initialize a dictionary to store correlation results
-    correlation_results = {}
-    p_value_results = {}
-    # List of your independent variables
-    all_signals = INDIVIDUAL_CHANS
-    df = dataframe_from_signals(signals_dict)
-
-    # Calculate correlation for each signal
-    # f = pearson3
-    f = spearmanr
-    for signal in all_signals:
-        for stat in statistics:
-            corr, pvalue = f(df[signal], df[stat])
-
-            # Store results in the dictionary
-            c = correlation_results.get(signal, {})
-            p = p_value_results.get(signal, {})
-            c[stat] = corr
-            p[stat] = pvalue
-            correlation_results[signal] = c
-            p_value_results[signal] = p
-
-    return pd.DataFrame(correlation_results), pd.DataFrame(p_value_results)
 
 if __name__ == '__main__':
     root_dir = os.path.join(os.getcwd(), "results")
